flatten: Ensure format is portable.  Change 'xferInt' to 'xferInt32', etc.
In order to ensure the serialization format is portable, arrange to
always use particular sizes, and big endian encoding.  This is an
incompatible change, so I renamed some affected methods to ensure
breakage and review.

+++ bflatten.h
@@ -79,7 +79,7 @@ public:
 
   // Flatten funcs
   virtual bool reading() const override { return m_stream.readMode(); }
-  virtual void xferSimple(void *var, unsigned len) override;
+  virtual void xferSimple(void *var, size_t len) override;
 };
